requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
astroid==3.3.8
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dill==0.3.9
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
importlib_metadata==8.6.1
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
npx==0.1.6
numpy==2.2.5
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyJWT==2.10.1
pylint==3.3.3
pymongo==4.12.0
pyparsing==3.2.3
PyPDF2==3.0.1
pystray==0.19.5
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
requests==2.32.3
rsa==4.9.1
setuptools==80.3.0
six==1.17.0
tomlkit==0.13.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
